bo4 fastfile dump pools

+++ src/acts/tools/ff/handlers/bo4/bo4_unlinker_rawfile.cpp
@@ -22,6 +22,7 @@ namespace {
 
 			std::filesystem::create_directories(outFile.parent_path());
 			LOG_INFO("Dump raw file {} 0x{:x}", outFile.string(), asset->len);
+			if (!asset->len && std::filesystem::exists(outFile)) return; // ignore empty files
 			if (!utils::WriteFile(outFile, asset->buffer, asset->len)) {
 				LOG_ERROR("Error when dumping {}", outFile.string());
 			}
